tests/wpt/tests/css/css-gaps/animation/gap-decorations-width-neutral-keyframe-003.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>gap decorations row-rule-width neutral keyframe</title>
    <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#column-row-rule-width">
    <meta name="assert" content="gap decorations row-rule-width with repeaters supports neutral keyframe.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/support/interpolation-testcommon.js"></script>
  </head>
  <body>
    <div id="target"></div>
    <script>
      test(() => {
        target.style.rowRuleStyle = 'solid';
        target.style.rowRuleWidth = 'repeat(2, 20px 20px) repeat(1, 30px)';
        var animation = target.animate([{}, {rowRuleWidth: 'repeat(2, 40px 40px) repeat(1, 40px)'}], 1000);
        animation.pause();
        animation.currentTime = 500;
        assert_equals(getComputedStyle(target).rowRuleWidth, 'repeat(2, 30px 30px) repeat(1, 35px)');
      }, 'gap decorations row-rule-width with repeaters supports neutral keyframe.');
    </script>
  </body>
</html>
